refactor: clean up imports and improve error handling in services and controllers

+++ src/main/java/friutrodez/backendtourneecommercial/service/ItineraireService.java
@@ -18,9 +18,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Stream;
+import java.util.Optional;
 
 /**
  * Service de gestion des itinéraires.
@@ -155,9 +154,14 @@ public class ItineraireService {
      */
     @Transactional
     public void deleteItineraire(long itineraireId, Utilisateur user) {
-        appartientRepository.deleteAppartientByIdEmbedded_Itineraire_UtilisateurAndIdEmbedded_Itineraire(
-                user, itineraireRepository.findById(itineraireId).get());
-        itineraireRepository.deleteById(itineraireId);
+        Optional<Itineraire> itineraire = itineraireRepository.findById(itineraireId);
+        if (itineraire.isPresent()) {
+            appartientRepository.deleteAppartientByIdEmbedded_Itineraire_UtilisateurAndIdEmbedded_Itineraire(
+                    user, itineraire.get());
+            itineraireRepository.deleteById(itineraireId);
+        } else {
+            throw new DonneesInvalidesException("L'itinéraire n'existe pas.");
+        }
     }
 
     /**
@@ -192,9 +196,8 @@ public class ItineraireService {
      *
      * @param itineraire l'itinéraire à lier
      * @param ids        un tableau ordonné des identifiants des clients
-     * @return un tableau des liaisons créées
      */
-    private List<Appartient> saveAppartientsFromListIdClients(Itineraire itineraire, String[] ids) {
+    private void saveAppartientsFromListIdClients(Itineraire itineraire, String[] ids) {
         List<Appartient> appartients = new ArrayList<>(ids.length);
 
         for (int position = 0; position < ids.length; position++) {
@@ -202,7 +205,7 @@ public class ItineraireService {
             appartients.add(new Appartient(new AppartientKey(itineraire, idClient), position));
         }
 
-        return appartientRepository.saveAll(appartients);
+        appartientRepository.saveAll(appartients);
     }
 
     /**
